# api/openapi.yaml
openapi: 3.1.0
info:
  title: Click & Collect API
  version: 0.1.0
  summary: Backend for the MegaLoja Online Click & Collect MVP.
  description: |
    REST API for creating orders with either standard delivery or
    pickup-in-store, listing stores, transitioning order status, and
    retrieving order notifications.

    All non-health endpoints are versioned under `/api/v1`.
  contact:
    name: click-and-collect
    url: https://github.com/gab-mello/click-and-collect
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/carts/{id}:
    get:
      tags: [carts]
      operationId: get-cart
      summary: Get a cart by ID (items hydrated)
      parameters:
        - $ref: "#/components/parameters/CartID"
      responses:
        "200":
          description: Cart found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404":
          description: Cart not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                not_found:
                  value: { error: "cart not found", code: "cart_not_found" }
  /api/v1/carts/{id}/items:
    post:
      tags: [carts]
      operationId: add-cart-item
      summary: Add (or increase quantity of) an item in the cart
      description: |
        Upserts the `(cart_id, product_id)` pair. If the item already exists,
        `quantity` is **added** to the current value. Returns the hydrated cart.

        The cart must be `ACTIVE`. The product must exist and be `active`.
      parameters:
        - $ref: "#/components/parameters/CartID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCartItemRequest" }
            examples:
              add_two:
                value: { product_id: prod-headphones, quantity: 2 }
      responses:
        "200":
          description: Item upserted. Cart returned with hydrated items.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400":
          description: |
            Validation error.

            - `invalid_body`
            - `invalid_quantity` — `quantity` must be greater than zero.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: |
            Cart or product not found. The `code` distinguishes:

            - `cart_not_found`
            - `product_not_found`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            Business rule violation.

            - `cart_not_active` — cart has already been checked out.
            - `product_inactive` — product exists but is not for sale.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/carts/{id}/items/{product_id}:
    patch:
      tags: [carts]
      operationId: update-cart-item-quantity
      summary: Set the quantity of an existing cart item
      description: Overwrites the quantity (does not add to it). The cart must be `ACTIVE` and the line must exist.
      parameters:
        - $ref: "#/components/parameters/CartID"
        - $ref: "#/components/parameters/CartItemProductID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCartItemRequest" }
            examples:
              set_to_two:
                value: { quantity: 2 }
      responses:
        "200":
          description: Item updated. Cart returned with hydrated items.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400":
          description: |
            - `invalid_body`
            - `invalid_quantity`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: |
            - `cart_not_found`
            - `cart_item_not_found`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            - `cart_not_active`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [carts]
      operationId: remove-cart-item
      summary: Remove a line from the cart
      parameters:
        - $ref: "#/components/parameters/CartID"
        - $ref: "#/components/parameters/CartItemProductID"
      responses:
        "200":
          description: Item removed. Cart returned with the remaining items.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404":
          description: |
            - `cart_not_found`
            - `cart_item_not_found`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            - `cart_not_active`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/carts/{id}/checkout:
    post:
      tags: [carts]
      operationId: checkout-cart
      summary: Convert an active cart into an order
      description: |
        Transactional cart checkout — the canonical way to create an order
        with items.

        Steps performed inside a single transaction:

        1. Lock the cart row.
        2. Verify the cart is `ACTIVE` and has at least one item.
        3. Re-read every product; reject inactive ones.
        4. Snapshot product `name` and `unit_price_cents` into `order_items` —
           future product price/name changes will not retroactively affect
           past orders.
        5. Compute `total_amount_cents` server-side.
        6. Apply the same delivery-method / pickup-store validation as
           `POST /api/v1/orders` (pickup orders require an active store and
           get a 6-digit `pickup_code`).
        7. Mark the cart `CHECKED_OUT`.

        Any failure rolls back; the cart remains `ACTIVE` and no partial
        order is persisted.
      parameters:
        - $ref: "#/components/parameters/CartID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutRequest" }
            examples:
              standard:
                summary: Standard delivery
                value:
                  customer_name: Grace Hopper
                  customer_email: grace@example.com
                  delivery_method: standard
              pickup:
                summary: Pickup in store
                value:
                  customer_name: Grace Hopper
                  customer_email: grace@example.com
                  delivery_method: pickup_in_store
                  pickup_store_id: store-sp-paulista
      responses:
        "201":
          description: Cart was checked out. Returns the created order (and a notification slot for symmetry with other endpoints — always null on checkout because the order starts in `AWAITING_PREPARATION`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutResponse" }
        "400":
          description: |
            Validation error.

            - `invalid_body`
            - `customer_name_required`
            - `customer_email_required`
            - `invalid_delivery_method`
            - `pickup_store_required`
            - `pickup_store_not_allowed`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: |
            - `cart_not_found`
            - `store_not_found`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            Business rule violation.

            - `cart_not_active`
            - `cart_empty`
            - `product_inactive`
            - `store_inactive`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
...
    CartID:
      name: id
      in: path
      required: true
      description: Cart identifier (16 lowercase hex chars).
      schema:
        type: string
        pattern: "^[0-9a-f]{16}$"
        examples: [bc3458adb995135f]
    CartItemProductID:
      name: product_id
      in: path
      required: true
      description: Product identifier of the cart line to act on.
      schema:
        type: string
        examples: [prod-headphones]
  schemas:
...
    CartStatus:
      type: string
      enum: [ACTIVE, CHECKED_OUT]
      description: |
        - `ACTIVE` — items may be added, updated, or removed; the cart can be checked out.
        - `CHECKED_OUT` — the cart was converted into an order and is now immutable.
    CartItem:
      type: object
      required: [id, cart_id, product_id, quantity, created_at, updated_at]
      properties:
        id:         { type: string, pattern: "^[0-9a-f]{16}$" }
        cart_id:    { type: string, pattern: "^[0-9a-f]{16}$" }
        product_id: { type: string, examples: [prod-headphones] }
        quantity:   { type: integer, minimum: 1, examples: [2] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Cart:
      type: object
      required: [id, status, items, created_at, updated_at]
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{16}$"
          examples: [bc3458adb995135f]
        status:
          $ref: "#/components/schemas/CartStatus"
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AddCartItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string, examples: [prod-headphones] }
        quantity:   { type: integer, minimum: 1, examples: [2] }
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1, examples: [2] }
    CheckoutRequest:
      type: object
      required: [customer_name, customer_email, delivery_method]
      properties:
        customer_name:  { type: string, examples: [Grace Hopper] }
        customer_email: { type: string, format: email, examples: [grace@example.com] }
        delivery_method:
          $ref: "#/components/schemas/DeliveryMethod"
        pickup_store_id:
          type: string
          description: Required when `delivery_method` is `pickup_in_store`; must be omitted otherwise.
          examples: [store-sp-paulista]
    CheckoutResponse:
      type: object
      required: [order]
      properties:
        order:
          $ref: "#/components/schemas/Order"
        notification:
          $ref: "#/components/schemas/Notification"
          description: |
            Reserved for symmetry with other endpoints. Always omitted on
            checkout because new orders start in `AWAITING_PREPARATION` and
            notifications are only emitted on the transition to
            `READY_FOR_PICKUP`.
    DeliveryMethod:
      type: string
      enum: [standard, pickup_in_store]
      description: |
        - `standard` — shipping; no store involvement.
        - `pickup_in_store` — customer picks up at `pickup_store_id`; server issues a `pickup_code`.
    Status:
      type: string
      enum: [AWAITING_PREPARATION, READY_FOR_PICKUP, COLLECTED, CANCELLED]
      description: Order lifecycle state. See `update-order-status` for allowed transitions.
    Order:
      type: object
      required:
        - id
        - customer_name
        - customer_email
        - delivery_method
        - status
        - total_amount_cents
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{16}$"
          examples: [9f1c2a4b8d6e7012]
        customer_name:
          type: string
          examples: [Grace Hopper]
        customer_email:
          type: string
          format: email
          examples: [grace@example.com]
        delivery_method:
          $ref: "#/components/schemas/DeliveryMethod"
        pickup_store_id:
          type: [string, "null"]
          description: Set only when `delivery_method` is `pickup_in_store`.
          examples: [store-sp-paulista]
        pickup_code:
          type: [string, "null"]
          description: 6-digit code issued by the server for pickup-in-store orders.
          pattern: "^[0-9]{6}$"
          examples: ["428193"]
        status:
          $ref: "#/components/schemas/Status"
        cart_id:
          type: [string, "null"]
          description: |
            Set when the order was created via `POST /api/v1/carts/{id}/checkout`.
            Null for orders created through the legacy `POST /api/v1/orders` path.
          pattern: "^[0-9a-f]{16}$"
          examples: [bc3458adb995135f]
        total_amount_cents:
          type: integer
          format: int64
          minimum: 0
          description: |
            Server-computed total over `items`. Always `0` for legacy orders
            created via `POST /api/v1/orders` (which have no items).
          examples: [129700]
        items:
          type: array
          description: |
            Snapshots of the products at checkout time. Empty array `[]` for
            legacy orders. Each item carries the product name and unit price
            captured at the moment the order was created, so historical orders
            are unaffected by later product changes.
          items: { $ref: "#/components/schemas/OrderItem" }
        created_at:
          type: string
          format: date-time
          examples: [2026-06-14T18:30:00Z]
        updated_at:
          type: string
          format: date-time
          examples: [2026-06-14T18:30:00Z]
    OrderItem:
      type: object
      required: [id, order_id, product_id, product_name, unit_price_cents, quantity, total_price_cents]
      properties:
        id:                 { type: string, pattern: "^[0-9a-f]{16}$" }
        order_id:           { type: string, pattern: "^[0-9a-f]{16}$" }
        product_id:         { type: string, examples: [prod-headphones] }
        product_name:       { type: string, examples: [Wireless Headphones X1] }
        unit_price_cents:
          type: integer
          format: int64
          minimum: 0
          examples: [59900]
        quantity:           { type: integer, minimum: 1, examples: [2] }
        total_price_cents:
          type: integer
          format: int64
          minimum: 0
          description: "Equals `unit_price_cents * quantity`."
          examples: [119800]
...
    Notification:
      type: object
      required: [id, order_id, message, created_at]
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{16}$"
          examples: [0a1b2c3d4e5f6071]
        order_id:
          type: string
          pattern: "^[0-9a-f]{16}$"
          examples: [9f1c2a4b8d6e7012]
        message:
          type: string
          examples: ["Order 9f1c2a4b8d6e7012 is ready for pickup. Use code 428193."]
        created_at:
          type: string
          format: date-time
          examples: [2026-06-14T19:05:12Z]
...
    Error:
      type: object
      required: [error]
      description: Error envelope used by every non-2xx response.
      properties:
        error:
          type: string
          description: Human-readable message.
          examples: ["order not found"]
        code:
          type: string
          description: Stable machine-readable code; safe to switch on.
          examples: [order_not_found]